requirements.txt
torch>=2.0.0
torch-directml>=0.2.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
Pillow>=8.0.0
joblib>=1.0.0
psutil>=5.8.0
pandas>=1.3.0
seaborn>=0.11.0 